+++ components/AdvancedSearch.tsx
@@ -5,12 +5,6 @@ import React, { Fragment, useEffect, useMemo, useState } from "react";
 import { useSuspenseInfiniteQuery } from "@tanstack/react-query";
 import { infiniteSearchOptions } from "@/lib/hooks/useSearch";
 import { useInView } from "react-intersection-observer";
-import Container from "@/components/Container";
-import { MultiSelect } from "@/components/ui/multi-select";
-import { cardRarity, cardTypes, colors, formatLegalities } from "@/lib/primitives";
-import MultipleSelector from "@/components/ui/multiple-selector";
-import { Select, SelectContent, SelectGroup, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select";
-import { DualRangeSlider } from "@/components/ui/dual-range-slider";
 import { Spinner } from "@/components/ui/spinner";
 import { Card as CardType } from "@/lib/types/card";
 import CardContextMenu from "@/components/CardContextMenu";
@@ -18,14 +12,14 @@ import { Card } from "@/components/Card";
 import { buildQueryParts } from "@/lib/utils";
 
 export default function AdvancedSearch() {
-  const [selectedColors, setSelectedColors] = useQueryState("colors", parseAsArrayOf(parseAsString).withDefault([]));
-  const [selectedCardTypes, setSelectedCardTypes] = useQueryState("cardTypes", parseAsArrayOf(parseAsString).withDefault([]));
-  const [selectedRarity, setSelectedRarity] = useQueryState("rarity", parseAsString.withDefault(""));
-  const [power, setPower] = useQueryState("power", parseAsArrayOf(parseAsInteger).withDefault([0, 10]));
-  const [year, setYear] = useQueryState("year", parseAsArrayOf(parseAsInteger).withDefault([1993, 2024]));
-  const [price, setPrice] = useQueryState("price", parseAsArrayOf(parseAsInteger).withDefault([0, 800]));
-  const [legalities, setLegalities] = useQueryState("legalities", parseAsArrayOf(parseAsString).withDefault([]));
-  const [page, setPage] = useState(1);
+  const [selectedColors] = useQueryState("colors", parseAsArrayOf(parseAsString).withDefault([]));
+  const [selectedCardTypes] = useQueryState("cardTypes", parseAsArrayOf(parseAsString).withDefault([]));
+  const [selectedRarity] = useQueryState("rarity", parseAsString.withDefault(""));
+  const [power] = useQueryState("power", parseAsArrayOf(parseAsInteger).withDefault([0, 10]));
+  const [year] = useQueryState("year", parseAsArrayOf(parseAsInteger).withDefault([1993, 2024]));
+  const [price] = useQueryState("price", parseAsArrayOf(parseAsInteger).withDefault([0, 800]));
+  const [legalities] = useQueryState("legalities", parseAsArrayOf(parseAsString).withDefault([]));
+  const [page] = useState(1);
 
   const query = useMemo(() => buildQueryParts({
     selectedCardTypes, selectedColors, selectedRarity, power, year, price, legalities
@@ -41,118 +35,32 @@ export default function AdvancedSearch() {
     }
   }, [inView, fetchNextPage]);
 
-  return (<Container title={"Advanced Search"}>
-      <div className={"grid grid-cols-4 gap-x-2 pb-10"}>
-        <MultiSelect
-          options={colors}
-          onValueChange={setSelectedColors}
-          defaultValue={selectedColors}
-          placeholder="Select Colors"
-          variant="default"
-          animation={0}
-          maxCount={3}
-        />
-        <MultipleSelector
-          defaultOptions={cardTypes}
-          onChange={async (values) => {
-            const selectedCardTypes = values.map((value) => value.value);
-            await setSelectedCardTypes(selectedCardTypes);
-          }}
-          placeholder="Select Card Types"
-          emptyIndicator={<p className="text-center text-lg leading-10 text-gray-600 dark:text-gray-400">
-            no results found.
-          </p>}
-          groupBy="group"
-        />
-        <MultiSelect
-          options={formatLegalities}
-          onValueChange={setLegalities}
-          defaultValue={legalities}
-          placeholder="Select Legalities"
-          variant="default"
-          animation={0}
-          maxCount={3}
-        />
-        <Select onValueChange={setSelectedRarity} value={selectedRarity}>
-          <SelectTrigger className="h-full">
-            <SelectValue placeholder="Select a Rarity" />
-          </SelectTrigger>
-          <SelectContent>
-            <SelectGroup>
-              {cardRarity.map((rarity) => (<SelectItem key={rarity.value} value={rarity.value}>
-                  {rarity.label}
-                </SelectItem>))}
-            </SelectGroup>
-          </SelectContent>
-        </Select>
+  return (<div className="mt-4">
+    {isLoading && <Spinner />}
+    {isSuccess && query.length > 0 && (<>
+      <h2 className="text-center text-3xl font-semibold">
+        Search Results
+      </h2>
+      {data.pages.length > 0 ? (<div className={"grid grid-cols-6 gap-2"}>
+        {data.pages.map((page) => (<Fragment key={page.nextId}>
+          {page.map((card: CardType) => (<CardContextMenu card={card} key={card.id}>
+            <Card card={card} key={card.id} />
+          </CardContextMenu>))}
+        </Fragment>))}
+      </div>) : (<p>No results found</p>)}
+      <div className="flex justify-center">
+        <button
+          ref={ref}
+          className={"mx-auto"}
+          onClick={() => fetchNextPage()}
+          disabled={!hasNextPage || isFetchingNextPage}
+        >
+          {isFetchingNextPage ? (<Spinner />) : (!hasNextPage && "Nothing more to load")}
+        </button>
       </div>
-      <div className={"grid grid-cols-3 gap-x-2 gap-y-10"}>
-        <div>
-          <label htmlFor="year-slider">Year</label>
-          <DualRangeSlider
-            id="year-slider"
-            value={year}
-            min={1993}
-            max={2024}
-            step={1}
-            label={(value) => value}
-            onValueChange={(value) => setYear(value)}
-          />
-        </div>
-        <div>
-          <label htmlFor="price-slider">Price</label>
-          <DualRangeSlider
-            id="price-slider"
-            value={price}
-            min={0}
-            max={800}
-            step={5}
-            label={(value) => value}
-            onValueChange={(value) => setPrice(value)}
-          />
-        </div>
-        <div>
-          <label htmlFor="power-slider">Power</label>
-          <DualRangeSlider
-            id="power-slider"
-            value={power}
-            min={0}
-            max={10}
-            step={1}
-            label={(value) => value}
-            onValueChange={(value) => setPower(value)}
-            className={"p-y-5 w-full"}
-          />
-        </div>
+      <div>
+        {isFetching && !isFetchingNextPage ? "Background Updating..." : null}
       </div>
-
-      <div className="mt-4">
-        {isLoading && <Spinner />}
-        {isSuccess && query.length > 0 && (<>
-            <h2 className="text-center text-3xl font-semibold">
-              Search Results
-            </h2>
-            {data.pages.length > 0 ? (<div className={"grid grid-cols-6 gap-2"}>
-                {data.pages.map((page) => (<Fragment key={page.nextId}>
-                    {page.map((card: CardType) => (<CardContextMenu card={card} key={card.id}>
-                        <Card card={card} key={card.id} />
-                      </CardContextMenu>))}
-                  </Fragment>))}
-              </div>) : (<p>No results found</p>)}
-            <div className="flex justify-center">
-              <button
-                ref={ref}
-                className={"mx-auto"}
-                onClick={() => fetchNextPage()}
-                disabled={!hasNextPage || isFetchingNextPage}
-              >
-                {isFetchingNextPage ? (<Spinner />) : (!hasNextPage && "Nothing more to load")}
-              </button>
-            </div>
-            <div>
-              {isFetching && !isFetchingNextPage ? "Background Updating..." : null}
-            </div>
-          </>)}
-      </div>
-    </Container>);
+    </>)}
+  </div>);
 }
